Extract iOS review notification scheduling
